VERSION 1.0.0 major security issue fixed
Bump major version due to the importance of this fix.

The code for dropping the group permissions was wrong.
Dropping user permissions was not directly wrong,
but both were missing error processing.

This all has been fixed.

NOTE THAT THIS VERSION STILL IS NOT THOROUGHLY TESTED.
(I already use it on the majority of my machines, though.)