VERSION 1.1.0-dev escapes, suid propagation, etc.
':' is escaped as '\:', '\' can be escaped as '\\:', so

'a:b'  must be written as 'a\:b'
'a\b'  can  be written as 'a\\:b'
'a\:b' must be written as 'a\\:\:b'

suid propagation means, execed programs, which run as root,
now get the uid unchanged, so it can drop privileges as usual
for suid programs.

Also on OOPS the return value now is 126.

(This is an intermediate development version.)