VERSION 2.1.0: Flags ordered correctly
This is not for security, but it might change the meaning
of existing config, hence the minor version is incremented:

Before flags I and F were mixed, so you had to give F after I,
which is plain wrong, as flags are considered to be sorted
alphabetically.

This is not so bad.  suid complains if anything is in the
wrong order, so if you have used I, you might have to
rething again.