MAJOR BUGFIX: minmax-parsing works now
Most puzzling is, that this bug was already was in the very first commit.
So apparently I never tested it before, but I am pretty sure I did,
but this perhaps was before the first checkin.

Note that no-minmax (no arguments allowed) and minmax=- (any number of arguments) worked,
as I needed both.