VERSION 2.3.0: Important minmax bugfix
Minmax (in flags) did not work, so the entry always failed.
Now it should work properly.

Note that there was no security problem,
as you would have noticed that it does not work if you tried.

The only problem was, that if you were not able to restrict the argument
counts individually.  Only none (no minmax) or any (-) worked.