MAJOR SECURITY: :root:, :sh:, :bash: and /cmd
When using the :root: modifier the targeted binary is
executed with the configured UID as real UID and the effective UID
of suid, which usually is root (hence the name of the modifier).

Before, however, the target could belong to the configured UID,
hence changed by the configured UID, such that the configured UID
could become root.  Now it is checked against the effective UID of
suid (which usually is root) such that the target cannot be
provided by the configured UID anymore.

Note that this differs from the :suid: modifier, which executes
the target with the configured UID as effective UID and the
real UID of the calling user.  The checks are done against the
confgured UID, which is no problem, as user are allowed to
impersonate themself.

Here's a table how of the permission sets of the exece'd command
and how the target was checked previously and now:

Modif.	rUID	eUID	sUID	chkold	chknew
(none)	c	c	c	c	c
:root:	c	o	o	c!WRONG	o
:suid:	r	c	c	c	c

- Real, Effective, Saved are the Linux terminology
- Configured=suid.conf, Original=Effective UID of suid
- chkold is how previously the target was checked
- chknew is how it is checked now

Getting permissions right ain't easy!

This is why it is so important to have powerful tools like suid
to do it for you, such that you do not fall into the same traps
like this here.

Drastical changes in :bash: and :sh: to get them working:

:bash: and :sh: modifiers.  These did not work at all before.
Instead of prepending the commands, they appended themself
as arguments.  Also fexecve() did not work with them.
This all has been fixed.

Also suid now allows to run commands which are directly in the
root filesystem, something like `/init.sh`.

Note that none of these bugs above affected you, if you used
suid without :bash:, :sh: (which did not work at all), :root:
and sane commands and scripts belonging to root which did live
in some subdirectory.