Detect missing argv[0] (CVE-2021-4034)
Note that suid was not affected by CVE-2021-4034 (missing argv[0])

Today I learned, that argv[0] can be NULL.

If everything works regularily, `suid` supports being run with
no arguments at all.  In this case Usage is printed, and thanks to
the security first programming, argv[0] was not accessed there
(to not leak information), hence giving no chance for exploits here.

But things might be different or weird, see CVE-2021-4034.

argv[0] is now tested to be present, as suid might use this value
in some rare cases (where a NULL argv[0] would suppress the error
message, leaving only some "OOPS:" on the line, which is not what
we want).