SUIDPPID added
This serves the purpose to transport the PID of the calling process
to things called suid.

In future there might be intermediate processes needed to do some things,
hence the $PPID might not be the caller of suid.  But this PID might be
needed to processes to add them to namespaces etc.

Hence a stable variable to transport this information.